Testable handler builder with coercion
The newly added rehearser.handler/handler function basically wraps
reitit.ring/ring-handler, adding the coercion stuff among other
mangling that needs to be present, and taking the intended
application-specific stuff as parameters. The rehearser.http-service
ns was converted to use this, so it should be able to handle
timestamps now.

Fixes: #73 Coercion to/from Instant